Revert "Small refactor"
This reverts commit 66e6afac3cc8467c4ef51b519a6f2cb0838bba62.

+++ src/main/java/com/qa/persistence/domain/User.java
@@ -23,9 +23,10 @@ public class User {
 	@Column(length = 10)
 	private String dietryRequirements;
 	@Column(length = 10)
-	private long phoneNumber;
+	private int phoneNumber;
 	@Column(length = 10)
 	private int age;
+
 	
 	@JoinColumn(name ="userID", referencedColumnName = "userID")
 	@OneToMany(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
@@ -37,7 +38,7 @@ public class User {
 
 
 	public User(String userName, String dietryRequirements,
-			long phoneNumber, int age) {
+			int phoneNumber, int age) {
 		this.userName = userName;
 		this.dietryRequirements = dietryRequirements;
 		this.phoneNumber = phoneNumber;
@@ -60,11 +61,11 @@ public class User {
 		this.userName = userName;
 	}
 
-	public long getPhoneNumber() {
+	public int getPhoneNumber() {
 		return phoneNumber;
 	}
 
-	public void setPhoneNumber(long phoneNumber) {
+	public void setPhoneNumber(int phoneNumber) {
 		this.phoneNumber = phoneNumber;
 	}
